requirements.txt
fastapi
uvicorn[standard]
pydantic
pytest
httpx
pytest-asyncio
psycopg2-binary